requirements.txt
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
flask-swagger-ui==4.11.1
greenlet==1.1.3.post0
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
six==1.16.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
zipp==3.9.0

locust~=2.14.0